pyproject.toml
[tool.poetry]
name = "convert-pdf-protected-to-img"
version = "0.1.0"
description = ""
authors = ["ccirelli <christopher.cirelli@aig.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pandas = "^2.1.1"
PyPDF2 = "^3.0.1"
python-decouple = "^3.8"
pycryptodome = "^3.19.0"
pdf2image = "^1.16.3"
pypdfium2 = "^4.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
